operator: refator code from k8s_node.go to cilium_node.go
The code present in k8s_node was not specific to a k8s node. Thus, this
commit moved the code into a better appropriate file, cilium_node.go
